requirements.txt
flask
flask-cors
requests
PyJWT
pycryptodome